+++ src/views/sidebar/parts/edit/folder-details.tsx
@@ -3,9 +3,10 @@
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
+import { Container, Divider, Padding, Row, Text } from '@zextras/carbonio-design-system';
+import { t } from '@zextras/carbonio-shell-ui';
 import React, { FC } from 'react';
-import { Container, Row, Padding, Text, Divider } from '@zextras/carbonio-design-system';
-import { Folder, t } from '@zextras/carbonio-shell-ui';
+import type { Folder } from '../../../../carbonio-ui-commons/types/folder';
 
 const bytesToSize = (bytes: number): string => {
 	const sizes = ['Bytes', 'KB', 'MB', 'GB', 'TB'];
@@ -39,7 +40,7 @@ export const FolderDetails: FC<{ folder: Folder }> = ({ folder }) => (
 			</Row>
 			<Row orientation="vertical" width="33.33%" crossAlignment="flex-start">
 				<Text size="small" color="secondary">
-					Size{t('label.size', 'Size')}
+					{t('label.size', 'Size')}
 				</Text>
 				<Padding top="extrasmall" />
 				<Text>{bytesToSize(folder.s || 0)}</Text>